problems

+++ ft_putstr.c
@@ -6,7 +6,7 @@
 /*   By: dasalaza <dasalaza@student.42barcelona.c>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/19 18:20:36 by dasalaza          #+#    #+#             */
-/*   Updated: 2023/09/22 23:25:15 by dasalaza         ###   ########.fr       */
+/*   Updated: 2023/09/23 12:38:21 by dasalaza         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,6 +18,6 @@ int ft_putstr(va_list args)
     int     len;
 
     str = va_arg(args, char *);
-    len = ft_strlen(str);
+    len = (int) ft_strlen(str);
     return ((int) write(1, str, len));
 }
